+++ Jitex.Tests.Context/Caller.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Runtime.CompilerServices;
+﻿using System.Runtime.CompilerServices;
 
 namespace Jitex.Tests.Context
 {
