+++ Jitex/Utils/AppModules.cs
@@ -1,10 +1,10 @@
-﻿using Jitex.Utils.Comparer;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
+using Jitex.Utils.Comparer;
 
-namespace Jitex.JIT
+namespace Jitex.Utils
 {
     internal static class AppModules
     {
